Reorganize tests so they're easier to find
